fix: removed screens
